Stats

+++ database.sql
@@ -30,7 +30,7 @@ CREATE TABLE answers(
     question_id INT,
     title VARCHAR(255) NOT NULL,
     number_order INT,
-    count INT DEFAULT 0,
+    count BIGINT DEFAULT 0,
     FOREIGN KEY (question_id) REFERENCES questions (id) ON DELETE CASCADE
 );
